+++ src/functions/serialization.rs
@@ -1,3 +1,4 @@
+use super::utils;
 use rustc_middle::ty::TyCtxt;
 use rustc_span::source_map::SourceMap;
 use rustc_stable_hash::{FromStableHash, SipHasher128Hash, StableHasher, hashers::SipHasher128};
@@ -27,10 +28,10 @@ impl SerFunction {
     pub fn new(fun: super::Function, tcx: TyCtxt, src_map: &SourceMap) -> Self {
         let inst = fun.instance;
         let def_id = format_def_id(&inst);
-        let file = super::file_path(&inst);
+        let file = utils::file_path(&inst);
         let attrs: Vec<_> = fun.attrs.iter().map(|a| a.as_str().to_owned()).collect();
         // Though this is from body span, fn name and signature are included.
-        let func = super::source_code_with(fun.body.span, tcx, src_map);
+        let func = utils::source_code_with(fun.body.span, tcx, src_map);
         let callees: Vec<_> = fun.callees.iter().map(|x| Callee::new(x, tcx, src_map)).collect();
 
         // Hash
@@ -81,8 +82,8 @@ pub struct Callee {
 impl Callee {
     fn new(inst: &Instance, tcx: TyCtxt, src_map: &SourceMap) -> Self {
         let def_id = format_def_id(inst);
-        let file = super::file_path(inst);
-        let func = super::source_code_of_body(inst, tcx, src_map).unwrap_or_default();
+        let file = utils::file_path(inst);
+        let func = utils::source_code_of_body(inst, tcx, src_map).unwrap_or_default();
         Callee { def_id, file, func }
     }
 }
